Add a tool to manage the VPN CRL
The vault terraform provider does not have the resources to manage a certificate revocation list, necessary for this scenario were we have long lived certificates. We need an additional tool to handle this by talking directly to the vault server, which does have the CRL management capabilities.
The idea is to create a tool that can be started as a server with a small http API that can be interacted with to manage the CRL.