Add a hourly goroutine that forces update of the CRL
The CRL has an validity period. Vault automatically recreates the CRL, with
new validity times, when adding certificates to the CRL. But in our use case,
the certificates are long lived and the CRL could be not updated in quite some
time. Instead instructing the user to configure a big lease time for the CRL,
use a cront task to force hourly recreation (rotate)of it.